Skip writing idpairs table to index dataset for direct FK relationships.
