+++ service/src/main/resources/config/verily/sdd_refresh0323/entitygroup/device_person_occurrence.json
@@ -7,16 +7,10 @@
     "dataPointer": "omop_dataset",
     "relationshipMappings": {
       "occurrenceToCriteria": {
-        "foreignKeyAttribute": "device",
-        "idPairsTable": { "table": "device_exposure" },
-        "idPairsIdA": { "column": "device_exposure_id" },
-        "idPairsIdB":  { "column": "device_concept_id" }
+        "foreignKeyAttribute": "device"
       },
       "occurrenceToPrimary": {
-        "foreignKeyAttribute": "person_id",
-        "idPairsTable": { "table": "device_exposure" },
-        "idPairsIdA": { "column": "device_exposure_id" },
-        "idPairsIdB":  { "column": "person_id" }
+        "foreignKeyAttribute": "person_id"
       },
       "criteriaToPrimary": {
         "idPairsTable": { "table": "device_exposure" },
