+++ service/src/main/resources/config/verily/sdd_refresh0323/entitygroup/note_person_occurrence.json
@@ -7,16 +7,10 @@
     "dataPointer": "omop_dataset",
     "relationshipMappings": {
       "occurrenceToCriteria": {
-        "foreignKeyAttribute": "note",
-        "idPairsTable": { "table": "note" },
-        "idPairsIdA": { "column": "note_id" },
-        "idPairsIdB":  { "column": "note_type_concept_id" }
+        "foreignKeyAttribute": "note"
       },
       "occurrenceToPrimary": {
-        "foreignKeyAttribute": "person_id",
-        "idPairsTable": { "table": "note" },
-        "idPairsIdA": { "column": "note_id" },
-        "idPairsIdB":  { "column": "person_id" }
+        "foreignKeyAttribute": "person_id"
       },
       "criteriaToPrimary": {
         "idPairsTable": { "table": "note" },
