+++ service/src/main/resources/config/verily/sdd_refresh0323/entitygroup/visit_person_occurrence.json
@@ -7,16 +7,10 @@
     "dataPointer": "omop_dataset",
     "relationshipMappings": {
       "occurrenceToCriteria": {
-        "foreignKeyAttribute": "visit",
-        "idPairsTable": { "table": "visit_occurrence" },
-        "idPairsIdA": { "column": "visit_occurrence_id" },
-        "idPairsIdB":  { "column": "visit_concept_id" }
+        "foreignKeyAttribute": "visit"
       },
       "occurrenceToPrimary": {
-        "foreignKeyAttribute": "person_id",
-        "idPairsTable": { "table": "visit_occurrence" },
-        "idPairsIdA": { "column": "visit_occurrence_id" },
-        "idPairsIdB":  { "column": "person_id" }
+        "foreignKeyAttribute": "person_id"
       },
       "criteriaToPrimary": {
         "idPairsTable": { "table": "visit_occurrence" },
