--- service/src/main/resources/config/verily/sdd_refresh0323/sql/cpt4_procedureOccurrence.sql
@@ -1,7 +1,0 @@
-SELECT po.procedure_occurrence_id, pc.id AS cpt4_id
-FROM `victr-tanagra-test.sd_20230328.procedure_occurrence` AS po
-
-JOIN `verily-tanagra-dev.aou_static_prep_useast1.prep_cpt` AS pc
-ON pc.concept_id = po.procedure_source_concept_id
-
-WHERE pc.type = 'CPT4'
