+++ underlay/src/main/java/bio/terra/tanagra/underlay/EntityGroup.java
@@ -69,7 +69,7 @@ public abstract class EntityGroup {
                 entityGroup.getName(), relationship.getName()));
       }
       RelationshipMapping sourceMapping =
-          RelationshipMapping.fromSerialized(serializedSourceMapping, sourceDataPointer);
+          RelationshipMapping.fromSerialized(serializedSourceMapping, sourceDataPointer, relationship);
 
       DataPointer indexDataPointer =
           entityGroup.getMapping(Underlay.MappingType.INDEX).getDataPointer();
@@ -84,7 +84,7 @@ public abstract class EntityGroup {
                       .getIndexDataMapping()
                       .getRelationshipMappings()
                       .get(relationship.getName()),
-                  indexDataPointer);
+                  indexDataPointer, relationship);
 
       relationship.initialize(sourceMapping, indexMapping, entityGroup);
     }
